+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_select_empty_selection.dart
@@ -8,7 +8,8 @@ class MPTH2FileEditStateSelectEmptySelection extends MPTH2FileEditState
         MPTH2FileEditStateClearSelectionOnExitMixin,
         MPTH2FileEditStateGetSelectedElementsMixin,
         MPTH2FileEditStateMoveCanvasMixin,
-        MPTH2FileEditStateOptionsEditMixin {
+        MPTH2FileEditStateOptionsEditMixin,
+        MPTH2FileEditStateResetAreaBorderCtrlMetaCycleMixin {
   MPTH2FileEditStateSelectEmptySelection({
     required super.th2FileEditController,
   });