+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_select_non_empty_selection.dart
@@ -10,7 +10,8 @@ class MPTH2FileEditStateSelectNonEmptySelection extends MPTH2FileEditState
         MPTH2FileEditStateClearSelectionOnExitMixin,
         MPTH2FileEditStateGetSelectedElementsMixin,
         MPTH2FileEditStateMoveCanvasMixin,
-        MPTH2FileEditStateOptionsEditMixin {
+        MPTH2FileEditStateOptionsEditMixin,
+        MPTH2FileEditStateResetAreaBorderCtrlMetaCycleMixin {
   MPTH2FileEditStateSelectNonEmptySelection({
     required super.th2FileEditController,
   });